parsercachepurging: Increase keys TTL
Initially the keys had a TTL of 30 days, but
due to disk space issues (T167784) it had to
be reduced to 22 days.

We have re-freshed the parsercache hosts (T208383),
so we are ready to go back to 30 days TTL.
However, we want to do a small increase first, to make
sure everything works as expected.

Bug: T210992
Change-Id: Ic8cb0a3edd53a98694fcfc5df8d2526da991e02a