Add apt pinning for buster
With the introduction of debian buster we will move to using puppet 5.10
as provided in the buster main repo.  however we are still relient on
facter 2.4.  This patch updates the manifest to ensure facter and puppet
are pinned correctly.

This change will change the behaviour of pinning on stretch as currently
we pin to release = $facts['lsbdistcodename'] which is redundent

Also remove `puppet_major_version` from the puppet files.  the variable
is now only used in puppet master related modules, profiles and roles,
including standalon and openstack componets.

Change-Id: I90c0c402911b73bfda5c35a410e0e68ddf30369e