phabricator: allow http from deployment hosts on stand-by servers
follow-up to Iffb4830660446748 where we allow http connections
from deployment servers to phabricator servers in addition to
just caching servers to make it possible to test changes with
tools like apache-fast-test or curl from deploy1001.

Because the current puppet code only opens any firewall holes
if on the "active server" we need this extra hole to allow only
deployment_servers (not caches) on the non-active (stand-by) servers
phab1002 and phab2001.

This will allow testing staged changes and the stretch upgrade
before switching Varnish and without needing to send debug headers.

Bug: T190568
Change-Id: Ifefe30bfea8b5837bd4eabb3aaec4bef52a0accd