varnish/trafficserver: switch parsoid-tests backend, rename director
Switch the backend of the parsoid-test host names from
ruthenium (jessie) to scandium (stretch).

Should not be merged yet, Subbu can confirm when, but review of the
code would be nice.

Since the name of the director is "ruthenium" as well, rename that to
a neutral name. It's a bad habit that we name directors after the
host names of backends.

Bug: T201366
Change-Id: I8fa7654380afd30485b39e00dcad9386c4c107ba