wmcs: prefix all of our customs scripts with 'wmcs-'
When doing operations, a way to quickly identify available commands is by using
the tab auto-completion feature of the shell. If we prefix all of our custom
scripts with the same string, a couple of keystrokes will show all of our
available options, which may increase a bit our overview of common operations
on a server or a cluster. Also, having a consistent namespace for commands
seems a good idea itself.

Also, we inconsistently deployed scripts in several different filesystem paths,
while there is usually a specific place for this: /usr/local/[s]bin/.
So, while at the renaming, reallocate scripts. Also add .py extension to python
scripts to trigger CI linters.

I may understand people don't want some of the changes proposed by this patch,
because it may change some already established workflows. I agree with that
and I will have no problem if we abandon/revert/withdraw this change.

Example of changes:
 * in puppet:
	modules/openstack/files/mitaka/admin_scripts/cold-migrate --->
	modules/openstack/files/mitaka/admin_scripts/wmcs-cold-migrate.py
 * in the resulting filesystem:
	/root/cold-migrate --->
	/usr/local/sbin/wmcs-cold-migrate

This change requires a strong review of our documentation at wikitech. I will
do once this commit is merged.

Change-Id: Ibdc41abc37576a7d9e70f19b94cbdc742fbb3948
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>