Improve CI checks to ensure a basic catalogue compiles on all supported OS's
This patch adds additional spec tests to ensure the standard module compiles
on all supported operating systems.  The standard module includes
many other modules and is the default module if there is no match in
sipe.pp. as such it should provide for a high level of code coverage

The spec helper adds a  monkey patch to disable app_management.
puppet-rspec enables this by default if it is supported by puppet.
however when this setting is enabled `site` becomes a reserved word.

I have created a PR upstream in an effort to get a more elegant solution
https://github.com/rodjek/rspec-puppet/pull/742

Bug: T215275
Change-Id: Ia0545ca68bf3345822c0831d55f8ad88cdabc509