gerrit: add icinga https check for actual JSON content
Add an additional monitoring check for Gerrit that should
trigger if the URL /r/changes/?n=25&O=81 does not return 200
and is at least 10,000 bytes in size. Current size at time of
testing is 27,156 bytes.

We already have checks for HTTPS generally working and the Gerrit
process running but we did not have a check that looks at content
for a specific Gerrit URL.

Create a virtual host for the gerrit service name and attach the
check to that rather than to any of the physical servers.

Bug: T215033
Change-Id: I459b4be0ec67788b8a5781b3df5dd0badbed856b