icinga/parsoid: no monitoring notifications on test servers
A server with a role defined as only used for testing should
not have active monitoring notifications. Something that is only
a test should never be important enough to actively ping a lot of people.

We are not completely removing monitoring though, it can still be seen
in Icinga web UI, just no active IRC bot lines or emails are being generated.

Bug: T201366
Change-Id: I8c54952b5eb80612c52a2d5e464d1c6244e427fd