Introduce systemd::slice::all_users
This class is a generalization of the awesome work did by Arturo
for the toolforge bastion hosts. The idea is to have a configurable
and generic way to set the user-.slice unit on hosts used by multiple
users to limit their capabilities.

Apply also this new class to profile::toolforge::bastion, the origin
of the code that has been generalized.

Bug: T212824
Change-Id: I341d7a7090d4f5a33ebf90b96c82ae2a2ac92a02