dbstore.my.cnf: Make the slave IDEMPOTENT
This file is only used by dbstore1002 which will be decommissioned soon
as part of T210478
This host is in a very bad state and data is already very corrupted.
As it keeps crashing almost everyday, when it comes back it comes back
with broken replication, which means we have to spend lots of time
fixing it.

Making it idempotent so any replication error will be skipped.

Bug: T213670
Change-Id: Ieb6b0be12c5d58798a8f378390f35215f5a444af