toolforge: bastion cgroup limits need a big boost in resources
People cannot run a basic compile with current settings, and this will
be a hard stop on migrating folks off the old grid.  With these settings,
spikes in a compile should still work.  Blocking all use of swap might still
stop a lot of builds, but this might allow PyYaml to compile (currently, it doesn't).

Bug: T215434
Change-Id: I941509274555af4df43a5df69afaf00d80811504