wiki replicas: Remove reference to old comment fields
This makes the following user-visible changes, which should be
announced. The fields being removed have been mostly empty for some time
now anyway, so hopefully people have already migrated away from them.

* `archive`: Remove `ar_comment`
* `archive_userindex`: Remove `ar_comment`
* `filearchive`: Remove `fa_deleted_reason` and `fa_description`
* `filearchive_userindex`: Remove `fa_deleted_reason` and `fa_description`
* `image`: Remove `img_description`
* `ipblocks`: Remove `ipb_reason`
* `ipblocks_ipindex`: Remove `ipb_reason`
* `logging`: Remove `log_comment`
* `logging_logindex`: Remove `log_comment`
* `logging_userindex`: Remove `log_comment`
* `oldimage`: Remove `oi_description`
* `oldimage_userindex`: Remove `oi_description`
* `recentchanges`: Remove `rc_comment`
* `recentchanges_userindex`: Remove `rc_comment`
* `revision`: Remove `rev_comment`
* `revision_userindex`: Remove `rev_comment`

The `pt_reason` field in `protected_titles` isn't being removed in this
patch, but will go away once the column is dropped in Wikimedia
production.

This patch also updates the compat views for the above tables to do an
(INNER) JOIN with the `comment` table rather than a LEFT JOIN, and
simplifies the conditions to rely only on comment_text with no reference
to the old (and mostly now empty) `xx_comment` fields.

Bug: T212972
Change-Id: I03f24c2b8b38541049d37593f2c1a03695323cf3