create a new role/profile 'hmmp' to replace role(simplelamp)
A role to replace role(simplelamp) that allows migrating projects
one-by-one instead of having to convert all at once when
editing simplelamp.

The goal is to stop using the "apache" and "mysql" modules and use
the "httpd" and "mariadb" modules instead.

This started as a copy of simplelamp and then did that conversion.

Also added a system::role to make lint happy, slightly renamed the
parameter, one wasn't needed anymore etc.

Intention is that projects listed here can switch to it and new
users can also have a simplelamp again that uses the newer code.

https://tools.wmflabs.org/openstack-browser/puppetclass/role::simplelamp

Bug: T215662
Change-Id: I9d76623249dc62e2a776a4597f1419b0c2bbfd71