profile::mediawiki::php: allow to fine-tune request timeouts
On php7 we can't assume that set_time_limit() will honor wall clock
limits, so per  I8faf8f71, we're using Excimer for setting our
fine-tuned limits. While Excimer does honor wall-clock limits, it's not
able to interrupt a long-running C function (like a mysql query), so we
also need to add a hard limit at the FPM level.

For now, a single limit of 200 seconds on appservers and of 1 day on
jobrunners has been added, but we should explore being a bit more
fine-grained when it comes to jobrunners/videoscalers and possibly using
multiple fpm pools.

Change-Id: Id9e0924ef7114c8004553db92c32950a5bd7b705