logstash: move role::ls::eventlogging to profile::logstash::collector
role::logstash::eventlogging is currently applied only to logstash1007 which
is a single point of failure. To address this, let's apply the eventlogging
logstash kafka input and filter to all logstash collectors.

This moves the contents of role::logstash::eventlogging into
profile::logstash::collector so that these logs may be consumed from kafka
by all of the logstash collectors, with unique group_ids based on datacenter.

Bug: T213898
Change-Id: Ib4bca5b474e4f328ee98b6280e9373cf9329f326