systemd: user slice: pinning for libsystemd0 as well
If you try to install systemd without a pinning for libsystemd0 you can get:

[...]
The following packages have unmet dependencies:
 systemd : Depends: libsystemd0 (= 239-12~bpo9+1) but 232-25+deb9u8 is to be installed
E: Unable to correct problems, you have held broken packages.

While at it, use a more robust approach for installing the correct version of
the package which is telling puppet exactly which version we want.
This will require us to manually update the version number in puppet in case
there is an update in the Debian side.

Also, consolidate package list to a single array.

Bug: T215154
Change-Id: Id9fe8605fbe93aec84577deb7b549622457e0e1c
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>