prometheus: don't require Prometheus::Server when writing k8s token
The dependency causes the token file to never be written, and that causes prometheus::server
resource to fail thus puppet never actually recovers.

Remove the dependency to give the token file to be written despite the fact that prometheus::server
will fail to find the token file on first puppet run.

Not the most elegant solution but alternatives I can think of would require e.g. esplicitly creating
all path components, with the right ownership (prometheus) which in turn requires the prometheus
package to be installed, and the package itself is installed by prometheus::server.

Bug: T187987
Change-Id: I781f9dc8047a7e9c89e5a8b35f1acb0f30809db4