wmcs services: introduce profile for maintain_dbusers in services nodes
The maintain_dbusers feature is now running in clouddb-services instead of physical
labstore servers.

The code was generated and adjusted from the puppet code used in the previous
setup (which we cannot drop yet). Differences can be checked with:

* diff -u modules/role/manifests/labs/db/maintain_dbusers.pp \
	  modules/profile/manifests/wmcs/services/maintain_dbusers.pp

* diff -u modules/role/templates/initscripts/labs/db/maintain-dbusers.systemd.erb \
	  modules/profile/templates/initscripts/wmcs/services/wmcs-maintain-dbusers.systemd.erb

The script itself not change, other than the name to include the wmcs- prefix.

Bug: T216373
Change-Id: Ia59639cc9a7831f8535b09ef6c9d42e42b9cd88d
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>