contint: phaseout android slave
The Jenkins job apps-android-wikipedia-periodic-test relied on the
Android SDK to be installed and access to kvm in order to spawn the
Android emulator.

The job is no more used and has been removed. Thus cleanup the role used
to provision that specific kind of slaves (they all have been deleted).

Bug: T198495
Change-Id: I6664f68ba2e5c82eae1507824beae280d5a958d4