mcrouter: allow async foreign set/delete WAN cache operations
There is no reason to have MediaWiki block on the cross-DC roundtrips,
as it has been discussed by SRE/Performance teams. MediaWiki is already
able to deal with failures in replicating to the non active DC, plus there
is no requirement up to now for strong consistency between active and standby
memcached shards.

This change introduces a new parameter to the wancache class to set async
routes for the non active DC. It will be used to incrementally roll out the
change and study mcrouter's behavior, rather than enabling the new settings
to all the mw hosts at once.

Change-Id: I9fd9c43a46f4949b63f141af586defed6b0ae95d