dumpsdistribution: Make pages go off for disk space
This is unlikely to page for disk space anytime soon, but it certainly
should.  I am unsure about the profile::base::notifications: critical portion.
If either goes down, this will disastrously impact toolforge or the web service
for dumps without an admin intervening since there is no automatic failover.
I gather from some of the commits that this will ensure it will page if it
goes down.

Bug: T217068
Change-Id: I3383700596274bedb8e474655bec999057a4c7f4