network: Fix list of bastions for labs realm
User:Addshore discovered that he wasn't able to connect to certain hosts via
the secondary bastion.
I checked the list and found the correct IP for bastion-eqiad1-02 was not listed -
in it's place was an IP for a random instance in the tools project.

While we're here, also remove a nonexistent restricted bastion entry.

Change-Id: Ie796d1a3f31bc40b5be8680aecb3ebd72c31efc9