Find db host and port using refinery
Production database replica are now sharded.  We have utilities that
grab configuration from mediawiki-config and find the appropriate shards
for a given database name.  These are deployed with refinery to all the
analytics machines.  We have moved the labs host to the same utilities
for consistency.  We might consider creating DNS locator records for the
labs host to match the production replicas.

Bug: T215290
Change-Id: Ifd3e0d06e358ffe6495f39c90ae79d92caeed19c