Create /etc/debdeploy-autorestarts.conf which lists all automated restarts
Create /etc/debdeploy-autorestarts.conf which lists all services which
get automatically restarted. This will be complemented by a patch on
the debdeploy side which trims the output in query_restart for any service
automatically restarted (right after a deploy, the restarts haven't kicked
in, but it's useful to be able to filter them out to review which services
are in need of a manual restart).

Change-Id: I4562255e4f94eb7310afe892cad5185c89c86ca6