scap: do not manage package versions via puppet by default
In production, we don't want to have puppet managing the package version
for debs - we have better tooling to handle software upgrades, it
interferes with server bootstrap if the upload of the package happens 
before the time when we're ready to deploy the new version, and overall 
is a bad idea.

After this patched is merged, debdeploy should be used for scap upgrades.

This patch adds a scap::version to the labs common file. This means that
the current behaviour is preserved for cloud VPS servers (so beta,
toolsforge and CI).

Bug: T217287
Change-Id: I4163dcb80519bcb522fc7a48721497a25aaefa88