Create /etc/debdeploy-autorestarts.conf which lists all automated restarts
Create /etc/debdeploy-autorestarts.conf which lists all services which
get automatically restarted. This will be complemented by a patch on
the debdeploy side which trims the output in query_restart for any service
automatically restarted (right after a deploy, the restarts haven't kicked
in, but it's useful to be able to filter them out to review which services
are in need of a manual restart).

The first version merged was reverted, the "absent" case missed the
mandatory "path" setting and given that the auto restart for "rasdaemon"
is absented on every non-buster system, it triggered puppet failures.

Change-Id: I46e17863fea497272ce7451bd8afee6ba49b61df