wmcs: Add profiles for oidentd proxy and client modes
Introduce profiles for provisioning the oidentd ident service as
a reverse proxy to be run on a NAT gateway host and as a client of
a proxy to be run on a CLoud VPS instance. This combination, along with
opening port 113 (auth / ident) on the NAT gateway, will allow Freenode
services to query for the local shell user identity of a connection.
This in turn will make it easier for Freenode to maintain spam blocks
while still allowing multiple Wikimedia IRC bots to run behind a small
number of NAT addresses.

The proxy mode service will be provisioned on the cloudnet* servers
which act as the public routers to our OpenStack Neutron SDN. The client
mode service will be provisioned on Son of Grid Engine exec nodes within
the Toolforge project. Other Cloud VPS projects may also choose to add
the client as needed.

The pidentd package that was previously deployed via
profile::toolforge::grid::exec_environ is removed. Ident protocol
responses are required by proxylistener which is used for registering
webservices on grid engine hosts.

Bug: T151704
Change-Id: I977bc62c6f206a959dfddb275526bddc11b60d6f