gerrit: Disable jgit gc
Hosts: cobalt.wikimedia.org, gerrit2001.wikimedia.org

Gerrit is currently affected by a bug in jGit wherein if a packfile on
disk does not match the packfile in the jGit in-memory index (for
instance in the situation where jGit runs a git gc and loads the
resulting packfile in memory, but the repo on disk is updated during
that process), rather than sensibly reload the packfile from disk, jGit
begins to throw errors until a(n as-yet-unknown) condition is met
causing it to reload the packfile from disk.

As we are only running git gc (via jGit) weekly, we have yet to hit this
condition. The issue is that since the Gerrit user database is now
stored in git, the inability of jGit to find a particular commit could
lead to users being locked out of the system.

This change trades off the performance and space gains we get from
running git gc for stability of the system. This change should be
reverted once a version of Gerrit is released with a fix for jGit.

Bug: T217497
Change-Id: I292880a38f5e3899d6ec35238cab68e7b1260240