Send traffic for citoid to kubernetes hosts as well
Add the kubernetes hosts. Note that the default values for pooled in
this service is "no", so we will have to manually set weight and status
which is exactly what we want.

Also instantiate the citoid service under the kubernetes cluster and add
the hosts as well. This is a noop currently as that cluster is not going
to receive traffic, but is a prerequisite for the next patch

Bug: T213194
Change-Id: Id06121eb649ef03b52c139859635ef1da2763aa2