Switch app servers to component/php72
Use component/php72 in favour of thirdparty/php72 which imported various
debs unmodified from the sury.org repository.

The new component contains the php7.2 source package from that repository,
but the package (and all related extensions) are rebuild entirely within
stretch (and the packages of the component). The repository formerly used
contains a number of updated core libraries which clash with our existing
HHVM installation. If we migrate to a new PHP release we only need to
rebuild the php7.2 package, the extensions retain the module ABI and don't
need a rebuild.

This also updates the PHP 7.2 packages to 7.2.15. I've done some smoke tests
on a codfw app server, initially I'll upgrade the mwdebug servers for further
testing.

Bug: T216712
Change-Id: Id9f1935d656e1a3c3855b4a4e06aeb44ae9efb01