mysqld-prometheus-exporter: Change the default arguments for buster
Buster (or more correctly, the version of prometheus-mysqld-exporter
available for install on buster, 0.11) breaks the formatting of the
default options, making it failing to start.

Because nobody is currently using options other than the default
ones, make those the actual defaults rather than duplicating them
everywhere. They can still be overriden, but otherwise, detect
the os version and apply the right config on each one.

This should be a noop everywhere as there is currently no buster
databases.

Bug: T161296
Bug: T213546
Bug: T193224
Change-Id: Ie5d3edc403e877edeb575701173e7c7d1ed61489