dumps distribution: remove labstore1006 for failover
Due to an unexplained reboot of this server that appears to be due to
firmware or hardware malfunction, we are failing over all services to
labstore1007 for troubleshooting.  This should be revert-able for failback

Note that DNS changes will be forthcoming to fail over the website and
clients need to be unmounted manually to prevent load issues in toolforge.

Bug: T217473
Change-Id: I8d4858567eafd5e7be39d9337ad9761d1109ea6e