icinga: add notes URLs to various monitoring checks, part 1
We want to make the notes_url parameter a required parameter
so that whenever a new monitoring check gets added a URL to a
runbook has to be added.

It turns out it's not feasible to do that with a jenkins-bot
vote for new code without adding it to all existing cases.

So this is the start to add one to all existing checks.

Either using Wikitech URLs or Phabricator workboard or project URLs
where a more specific one is not available.

This at least helps users to navigate to where tickets should be filed
and by checking the members of a Phabricator project one can determine
which users should likely be pinged about it.

They can and should be replaced with more specific runbook URLs on
Wikitech if/once these are available.

Bug: T197873
Change-Id: I883b2c03c422255232e2ddcdb3b0409dcb284cf6