icinga/elasticsearch: add notes URLs
We want to make the notes_url parameter a required parameter so
that monitoring checks have a URL to some kind of runbook or
background information that users can read if an alert happens to
trigger.

This is an effort to add one to all existing checks before being
able to do so.

Bug: T197873
Change-Id: Idcbf1ed620697cc69fae4e32159319b6a53cce07