Assign role::analytics_cluster::superset to analytics-tool1004
This new VM runs Debian Buster (notably with Python 3.7). The plan
is to create another database on an-coord1001 called 'superset-prod'
(to differenciate it from superset-staging, avoiding mistakes) that
will be populated with a snapshot of the current 'superset' database.

Eventually the goal is to have only analytics-tool1004 running superset,
from the 'staging-prod' database.

Bug: T212243
Change-Id: If2ba33f10270db2e59dc77f36cc043396d02f299