icinga/restbase/eventbus: add notes URLs
We want to make the notes_url a required parameter so that
all Icinga checks have a link to some kind of runbook or background
info for a user who happens to see a triggered alert.

Ideally that URL contains info how to react to the alert.

Are these ok for now or are there better ones? Maybe Phabricator
workboards are an option as well to tell users where to create tickets.

Bug: T197873
Change-Id: I387de73a9ead590e108caa534d38f150a33e9fb2