icinga/toollabs: set notes URLs for toolforge related checks
We want to make the notes_url a required parameter so that
all Icinga checks have a link to some kind of runbook or background
info for a user who happens to see a triggered alert.

Ideally that URL contains info how to react to the alert.

Does it make sense to link to this Phabricator workboard?
Could there be better ones on Wikitech for each of the checks,
maybe anchors within the same page?

Bug: T197873
Change-Id: Iad7660a65bb5b011040396a9afd080db64fd75da