toolforge: Rewrite envelope From headers when relaying
Add a rewrite rule to the exim4 config used by the Toolforge smarthost
to convert envelope From addresses matching '*@*.tools.eqiad.wmflabs'
internal email addresses to 'root@tools.wmflabs.org'. This is similar to
the '*@*.eqiad.wmflabs' to 'root@wmflabs.org' rewrite that is done by
the generic Cloud VPS smarthost, but scoped to the Toolforge project.

Follow up to 983f0cf3 which made these same changes, but to a template
that is cruft from an older role that is no longer in use.

Bug: T213416
Change-Id: I51f65237da1a3366059f697e3de7a0cc9531d3cb