Raise memcached dedicated memory on mc1019
In T217731 it was noted that there is currently a lot of
free ram on the mc10XX hosts (not even used by page cache).
At the same time, memcached constantly evicts keys in a lot of
slabs to have room for newer ones, and this might have bad effects
on the overall hit-ratio (that is really good to be fair, but
it could be probably improved).

This test adds ~10G of memory to memcached to see what effects it
could have on evictions/hit-ratio.

Bug: T217731
Change-Id: Ia7d535d40de6d2a9610aa153e6bbeff614c7df44