toolforge: limit file size to 50 GB
A single runaway tool can effectively fill NFS with a log.  This has
happened many times.  In lieu of currently having filesystem quotas
this will limit the size of a file that a toolforge user process can
create to 50 GB, which is pretty large in and of itself.

There are presently zipped files already on the NFS that exceed this size.

Bug: T122508
Change-Id: Icc5bc553308590cecfdf8c01851b50161ae0f33a