profile::kafka::broker: move java.security settings to its own class
The effort made to establish a good set of settings for java.security
in Kafka should be re-usable across services that want to use TLS
following what we can call WMF guidelines. Hadoop is the next service
lined up for this config, so the idea is to create a java::security
class that could be re-used across classes.

The name might be too generic and can be discussed, this is only a change
to start a discussion among people interested (security and java service
owners mostly).

Bug: T217412
Change-Id: I37cb4108bd7466d044e9c1de8cea50341dd91648