systemd: slice for all users: fix apt pinning
By the time of this patch, systemd v239 is no longer present in the
stretch-backports repo. We need a more flexible pinning, since any version
higher than 239 should be good for us.

Using ensure => latest means that puppet will automatically update systemd when
a new package is present in the repo. That may have consequences as well.

While at it, fix the array joining. It needs to be space separated.

Bug: T215154
Change-Id: I01ade3836a698e9f65f609966a09aa874cad1f8a
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>