openstack: drop -t jessie-backports for openstack component installs in stretch
No longer required in Debian Stretch. Since jessie-backports is no longer
installed as a repo, this will even produce an apt error (referencing a repo
not present in the system).

The change is a bit more complex than it might look like at first, as there
are several questions involved:

 * Will the apt resolver properly install mitaka instead of newton (stretch)?

It turns out that apt will choose the right version of the packages because we
install a repo for the openstack-mitaka-jessie component from jessie-wikimedia,
which has a default apt pinning with 1001 priority, which is higher than other
repos (stretch and stretch-backports).

 * Why is this patch introducing new package declarations and relationships?

The new repository configuration, along with apt pinnings, and how we install
the packages (now without `-t jessie-backports`) has consequences on how the
resolver installs the packages. The new openstack-mitaka-jessie component has
higher apt pinning value, and that may result in some assumptions being no
longer valid. An example of this is virt-top. Previously, the resolver was smart
enough to install the package the right version of libvirt0 (a depends of
virt-top), but now if we don't specify ordering, apt will try downgrading it
and puppet agent will fail.

For the record, this patch has been tested to work for fresh installations of:
 * cloudnet servers in mitaka/stretch
 * cloudvirt servers in mitaka/stretch
(following the instructions at:)
 https://wikitech.wikimedia.org/wiki/Portal:Cloud_VPS/Admin/Puppet_testing

But will require testing in physical hardware as well.

Bug: T216497
Change-Id: I7be898ef18a5724a1eb312efa23d17d983b192f8
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>