Send traffic for cxserver to kubernetes hosts as well
Add the kubernetes hosts. Note that the default values for pooled in
this service is "no", so we will have to manually set weight and
status which is exactly what we want.

Also instantiate the cxserver service under the kubernetes cluster and add
the hosts as well. This is a noop currently as that cluster is not going
to receive traffic, but is a prerequisite for the next patch

Bug: T213195
Change-Id: Ie47d9457bda8952b4d6cfa208f48e670923fd5a6