Revert "openstack: nova: mitaka: stretch: install python-dogpile.core from jessie"
This is a revert of commit c104b14bb66850a33bafe5b1d8733b5110a6c74f, but
adapted to the amount of changes in the git repo in meantime.

We no longer detect any issue with the python-dogpile.core package. Moreover,
enabling the jessie repo in stretch causes some complexities we would rather
avoid.

This has been tested in the 2 main server flavors that are using mitaka/stretch
right now, which are cloudvirts and cloudnet servers, using the docs at:
 https://wikitech.wikimedia.org/wiki/Portal:Cloud_VPS/Admin/Puppet_testing

Bug: T216497
Change-Id: I179c16b81bc7e685a1fa7de6bea904f32d59bdea
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>