aptrepo: drop more packages from openstack-mitaka-jessie
We don't need them here.

In fact, this may create resolver issues when using this repo in stretch. This
is because we install the openstack-mitaka-jessie repo with a high apt pinning
priority, which means the resolver uses all the packages in this repo as
candidates, which may result in weird clashes when combining with stretch
packages.

Bug: T216497
Change-Id: I615e98d1e74e6e7383fc4cdb2fff5bb57664bc68
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>