ATS: disable max_doc_size
Setting max_doc_size in records.config as we did in a9d5c83 has the
undesirable side-effect of disabling Read While Writer behavior.

We do have Lua code in place to make sure that large files are not
cached. Explicitly set max_doc_size to 0 in records.config and add a
comment explaining why we are doing this.

Bug: T213263
Change-Id: I0a3fc152b0f2f2385b246ea81bbb282db9ea6e05