openldap: Enforce cn uniqueness
The way we use the cn LDAP attribute throughout the infrastructure
justifies enforce uniqueness of it, at least for the posixAccount
objectClass entries. Add that.

The database has been audited and there were only 2, disabled and
clearly malicious users that had the same cn. The cn was a description
of the disabling. Those were fixed manually

Change-Id: I54e71756247bc1abd2b0f33f5a237c3a48acc7c6