acme_chief: Add security::access::config on passive host in cloud
This is not particularly relevant in production where if a user exists on the
host, either through the admin module or otherwise, and has an SSH key, it can
log in.
However, in the labs realm, everyone has to pass one of the security rules, the
usual one of course being the one which disallows everyone not in the
project-$project group in LDAP.
See scap::target for another example of this.

Change-Id: Ibff79509bb3e1defadad0f74e103984e3b38b644