mariadb-backups: Make some dump options explicit
Rotate and compress was implicit on the default configuration.

Make those explicit, in particular, always rotate, and don't
consolidate per-database except on x1 and s3, and don't create
a .tar.gz for the whole backup.

Make the retention policy 18 days (keeping at least 1 good backup
in addition to the latest one).

Bug: T206203
Change-Id: Id9201d4088aba8b26e2e836abc931e9cc2a391dd