Move qualified parameters to there correct location
After speaking with joe it seems that common.yaml should only contain
unqualified variables.  If a default value is required in WMCS/OpenStack
it should be placed in `hieradata/labs.yaml`.  If it is for production
then it is put in a file under `hieradata/common/` such that a::b::c would be
placed in hierdata/common/a/b.yaml.

This CR moves all fully qualified variables to there correct location
and further add's the common_yaml rake check to ensure errors are not
introduced in the future.

Failed ci example:
 * https://integration.wikimedia.org/ci/job/operations-puppet-tests-stretch-docker/8236/console
Success ci example:
 * https://integration.wikimedia.org/ci/job/operations-puppet-tests-stretch-docker/8238/console

Hosts: cumin1001.eqiad.wmnet,netmon1002.wikimedia.org,labpuppetmaster1001.wikimedia.org

Change-Id: Iea34ccb8340257ddcab32a7b4588025d90eff23e