Add curl to standard packages
Up to stretch, curl was implicitly installed on all our servers as
debian-goodies was installed by default and it pulled in curl via a
Depends. In buster this is no longer true; debian-goodies now only
recommends curl (and we don't installed recommended packages
automatically).

As a fix, simply add curl to default packages, plenty of other scripts
(and people) expect it to be present. Also clean up a few existing
places were curl was installed before.

Bug: T213527
Change-Id: Idda3861356149e9007b14e81bb17a757c7bec40e