phd: restart on failures
Rate-limited auto-restart of Phabricator's 'phd' daemon that runs
background tasks.

Also remove an erronenous RemainAfterExit; per the systemd docs for
a type=forking service, the service will be considered 'running' as
long as at least one process remains in the cgroup after running the
start command.  Inspection of the phd cgroup on phab1001 using systemctl
shows all of its associated processes.  (This change is also necessary
for autorestart to work if/when phd terminates abnormally.)

Change-Id: Ib717aa98cdc35c62b55e75b4b0fc80f4db073414