striker: let uwsgi container and app logs flow to stdout/stderr
Remove special log routes to local file and remote logstash in favor of
default uWSGI logging behavior which emits messages to stdout/stderr.
Running under systemd means that stdout/stderr will be captured by
journald which will make the log events available for local inspection.
Normal journald configuration will also make it possible for the log
events to be mirrored to an rsyslog instance, but that is left for
another patch.

Bug: T217932
Depends-On: Id4f52f2d5b8ba920d2a65ee7ea671249b6c3d9a6
Change-Id: I40c84663fcb259680a5157959f266e5f6a9fa9c7