WMCS: introduce sssd, replacing nscd/nslcd
Introduce the sssd component to WMCS, replacing nscd/nslcd. This new
daemon is responsible for contacting LDAP servers and serving both nss and pam
for user account and logins.

By using this more modern piece of software, we expect to reduce the impact
Toolforge has on LDAP servers. Also, sssd is easier to debug, more
configurable and should help us have better control on what's going on for
user/groups/LDAP in Toolforge in general.

To be able to use this LDAP client stack, an explicit hiera key is required:
 profile::ldap::client::labs::client_stack: sssd

The default value is 'classic' which uses the old nscd/nslcd stack.

Bug: T218126
Change-Id: I1103a1a5cd366922cea4502f3328335cdf389810
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>