profile::hadoop::common: explicitly set if TLS keys are deployed or not
This change seems to be the cleanest solution to avoid having the TLS configuration
directives spread in multiple places (rather than in a single place in
common.yaml) and at the same time avoid deploying TLS keys and their
xml config where not needed. The first use case is the Analytics coordinator,
where mapred-site.xml needs to be updated to use the encrypted shuffle
(so reducers of hive jobs know how to properly fetch map data via https)
but there is no need for TLS keys and related trustore/keystore settings.

Since this part of the profile is in rapid change and evolution, a completely
automated deployment of TLS keys may happen in the future when all
the settings and tunables will be clear.

Bug: T217412
Change-Id: I569eae1b7ed963180f361b4d8be2f12b7ae4c1ce