sonofgridengine: link hostgroup processing to host processing
This resolves part of the issue with deleted hosts. Any sgeexec host
should be handled properly with this.  Where it is sure to fail still is
webgrid queues.  Those still manage the hosts via direct injection into
the queue descriptions.  Honestly, they probably should be changed to use
hostgroups and do less work in puppet as a result.  Then this would be enough.

Babysteps.

Bug: T216151
Change-Id: If1261161de85810646f3db4d063b06f90b7271fa