Add agetty to filter_services list of debdeploy
agetty spaws the ttys on the terminal console and for the serial console.
These can be restarted via getty@ttyX.service and serial-getty@ttySX.service
units shipped by systemd, but they are too disruptive to automatically
restart them; the restart quits the current terminal session, which would
be too disruptive (especially during hardware maintenance over the serial
console). In addition, the ttys and the serial console are only exposed to
a narrow set of already privileged users and the attack foot print is
quite small (it only links against basic glibc functions).

If we ever run into a severe vulnerability in glibc which is exposed to
agetty we'd coordinate reboots to effect the wider glibc update anyway.

As such, add agetty to the filter list of debdeploy.

Bug: T135991
Change-Id: Ife4403e73f8006de24690818171b2f73a47fe5e3