sonofgridengine: SGE_DEBUG_LEVEL required for sge shadow to run
At some point, and I really don't know why, sge_shadowd suddenly requires
a debug level to be set or it will not start in "ND" mode.  It could
have something to do with a return code about a localization message file.

The important thing is that this stops the bleeding.

Bug: T211055
Bug: T219187
Change-Id: Ia8455e26e010fb9053a0aee142db77e13f6ab26c