dynamicproxy: Prevent STS header from non-TLS connections
Per RFC 6797: "An HSTS Host MUST NOT include the STS header field in
HTTP responses conveyed over non-secure transport."

Also be explicit in the WMCS use of `ssl_ciphersuite` that we do not
want automatic management of STS headers. This is the default today, but
could change in the future.

Bug: T102367
Change-Id: I51f85adfb385e6357e27816ac513fed60ec2084f