mariadb-backups: Once transferred, chown and prepare them as non-root
Backups with snapshoting method require root access for transfer, as
it requires access to privileged data on the /srv/datadir directory.

However, once trasferred, we belive the rest of the operations doesn't
require high privileges, just full access to the data. Change the
script so that it chowns data to the dump user and group, and then
executes the prepare backup script as sudo dump.

Bug: T206203
Change-Id: I476858192dac06e6793a63240c03c95bf5f020df